refactor: 'MainTemplate 내 Weather 관련 리팩토링'
Related to: #40
